requirements (1).txt
streamlit
langchain
faiss-cpu
openai
google-generativeai
langchain-community
tiktoken
